Don't add children to new parent as the name depend on the parent DirID
This works only if we have a 1:1 mapping of source and target folder that means move(/foo/bar, /foo/baz) which is not the case using our CryptoDecorator, there the cipher text name depend on the parent dir so we cannot simply append the child to the new parent. At this layer we can only mark the target parent as dirty so it will be fetched from the cloud.