If parent not in cache, request it in CachedPropfind's itemMetadata()
If parent is not available in cache, request parent in CachedPropfindEntryProvider's itemMetadata to avoid listings against `/foo/bar/baz` while `/foo/bar/` has not yet been queried which can lead to spam requests as `isChildrenFetched` is in this case not set in the parent and would also not be set using a PROPFIND on the child. Therefore listing on the parent.